[WIP] Extract transform attrs to a separate class.
