+++ dali/operators/geometry/coord_transform.h
@@ -28,22 +28,10 @@ namespace dali {
 #define COORD_TRANSFORM_DIMS (1, 2, 3, 4, 5, 6)
 
 template <typename Backend>
-class CoordTransform : public Operator<Backend> {
+class CoordTransform : public Operator<Backend>, private MTTransformAttr {
  public:
-  explicit CoordTransform(const OpSpec &spec) : Operator<Backend>(spec) {
+  explicit CoordTransform(const OpSpec &spec) : Operator<Backend>(spec), MTTransformAttr(spec) {
     dtype_ = spec_.template GetArgument<DALIDataType>("dtype");
-    has_mt_ = spec_.HasArgument("MT");  // check if there's a regular argument MT (not input)
-    has_mt_input_ = spec_.HasTensorArgument("MT");  // check if there's an Argument Input MT
-    has_matrix_ = spec_.HasArgument("M");  // check if there's a regular argument M (not input)
-    has_matrix_input_ = spec_.HasTensorArgument("M");  // check if there's an Argument Input M
-    has_translation_ = spec_.HasArgument("T");  // ...and similarly, for T
-    has_translation_input_ = spec.HasTensorArgument("T");
-
-    bool has_fused_arg = has_mt_ || has_mt_input_;
-    bool has_separate_args = has_matrix_ || has_matrix_input_ ||
-                             has_translation_ || has_translation_input_;
-    DALI_ENFORCE(!(has_fused_arg && has_separate_args), "The combined ``MT`` argument cannot be "
-      "used together with separate ``M``, ``T`` arguments.");
   }
 
   bool CanInferOutputs() const override { return true; }
@@ -127,156 +115,6 @@ class CoordTransform : public Operator<Backend> {
       ProcessMatrixArg(ws, "M", N);
       ProcessTranslationArg(ws, "T", N);
     }
-  }
-
-  bool HasFusedMT() const {
-    return has_mt_ || has_mt_input_;
-  }
-
-  bool HasMatrixRegularArg() const {
-    return has_matrix_ || has_mt_;
-  }
-
-  bool HasMatrixInputArg() const {
-    return has_matrix_input_ || has_mt_input_;
-  }
-
-  void ProcessMatrixArg(const workspace_t<Backend> &ws, const char *name, int N) {
-    bool is_fused = HasFusedMT();
-    int cols = input_pt_dim_ + (is_fused ? 1 : 0);
-    if (is_fused)
-      translation_.clear();
-
-    if (HasMatrixRegularArg()) {
-      mtx_ = spec_.template GetRepeatedArgument<float>(name);
-      if (mtx_.size() == 1) {
-        output_pt_dim_ = input_pt_dim_;
-        mtx_.resize(output_pt_dim_ * input_pt_dim_);
-        FillDiag(mtx_, mtx_[0]);
-        if (is_fused) {
-          translation_.resize(output_pt_dim_, 0);
-        }
-      } else {
-        DALI_ENFORCE(mtx_.size() % cols == 0, make_string("Cannot form a matrix ",
-          mtx_.size(), " elements and ", cols, "columns"));
-        output_pt_dim_ = mtx_.size() / cols;
-        if (is_fused)
-          SplitFusedMT();
-      }
-      Repeat(per_sample_mtx_, mtx_, N);
-      if (is_fused)
-        Repeat(per_sample_translation_, translation_, N);
-    } else if (HasMatrixInputArg()) {
-      const auto &M_inp = ws.ArgumentInput(name);
-      auto M = view<const float>(M_inp);
-      DALI_ENFORCE(is_uniform(M.shape), "Matrices for all samples int the batch must have "
-        "the same shape.");
-      DALI_ENFORCE(M.shape.sample_dim() == 0 || M.shape.sample_dim() == 2, make_string(
-        "The parameter M must be a list of 2D matrices of same size or a list of scalars. Got: ",
-        M.shape));
-      if (M.shape.sample_dim() == 0) {  // we have a list of scalars - put them on diagonals
-        output_pt_dim_ = cols;
-        int mat_size = output_pt_dim_ * cols;
-        per_sample_mtx_.resize(mat_size * N);
-        for (int i = 0; i < N; i++) {
-          FillDiag(make_span(&per_sample_mtx_[i * mat_size], mat_size), M.data[i][0]);
-        }
-        translation_.resize(output_pt_dim_, 0);
-        Repeat(per_sample_translation_, translation_, N);
-      } else {
-        DALI_ENFORCE(M.shape[0][1] == cols, make_string("The shape of the argument ``", name,
-           "`` does not match the input shape. Got ", M.shape[0], " matrices and the input "
-           "requires matrices with ", cols, " columns"));
-        output_pt_dim_ = M.shape[0][0];
-        int mat_size = output_pt_dim_ * input_pt_dim_;
-        per_sample_mtx_.resize(mat_size * N);
-        if (is_fused)
-          per_sample_translation_.resize(output_pt_dim_ * N);
-        for (int s = 0; s < N; s++) {
-          for (int i = 0; i < output_pt_dim_; i++) {
-            for (int j = 0; j < input_pt_dim_; j++)
-                per_sample_mtx_[s * mat_size + i * input_pt_dim_ + j] = M.data[s][i * cols + j];
-            if (is_fused)
-              per_sample_translation_[s * output_pt_dim_ + i] = M.data[s][i * cols + cols - 1];
-          }
-        }
-      }
-    } else {
-      output_pt_dim_ = input_pt_dim_;
-      if (static_cast<int>(mtx_.size()) != output_pt_dim_ * input_pt_dim_) {
-        mtx_.resize(output_pt_dim_ * input_pt_dim_, 0);
-        FillDiag(mtx_, 1);
-        Repeat(per_sample_mtx_, mtx_, N);
-        if (is_fused) {
-          translation_.resize(output_pt_dim_, 0);
-          Repeat(per_sample_translation_, translation_, N);
-        }
-      }
-    }
-  }
-
-  void SplitFusedMT() {
-    int cols = input_pt_dim_ + 1;
-    translation_.resize(output_pt_dim_);
-    for (int i = 0; i < output_pt_dim_; i++) {
-      // store the last column in the translation vector
-      translation_[i] = mtx_[i * cols + input_pt_dim_];
-
-      // compact the matrix
-      if (i > 0) {
-        for (int j = 0; j < input_pt_dim_; j++)
-          mtx_[i * input_pt_dim_ + j] = mtx_[i * cols + j];
-      }  // else: the first row is already where it should be
-    }
-    mtx_.resize(output_pt_dim_ * input_pt_dim_);
-  }
-
-  /** @brief Fill the diagonal with a scalar value, put zeros elsewhere */
-  template <typename Container>
-  void FillDiag(Container &&mtx, float value) {
-    assert(static_cast<int>(size(mtx)) == input_pt_dim_ * output_pt_dim_);
-    for (int i = 0, k = 0; i < output_pt_dim_; i++)
-      for (int j = 0; j < input_pt_dim_; j++, k++)
-          mtx[k] = (i == j ? value : 0);
-  }
-
-  void ProcessTranslationArg(const workspace_t<Backend> &ws, const char *name, int N) {
-    if (has_translation_) {
-      GetSingleOrRepeatedArg(spec_, translation_, name, output_pt_dim_);
-      Repeat(per_sample_translation_, translation_, N);
-    } else if (has_translation_input_) {
-      const auto &T_inp = ws.ArgumentInput(name);
-      auto T = view<const float>(T_inp);
-      DALI_ENFORCE(is_uniform(T.shape), "Translation vectors for all samples must have "
-        "the same shape.");
-      DALI_ENFORCE(T.shape.sample_dim() == 0 || T.shape.sample_dim() == 1, "The translation "
-        "argument input must be a list of 1D tensors or a list of scalars.");
-      translation_.clear();
-      per_sample_translation_.resize(N * output_pt_dim_);
-      if (T.shape.sample_dim() == 0) {
-        for (int i = 0, k = 0; i < N; i++)
-          for (int j = 0; j < output_pt_dim_; j++, k++)
-            per_sample_translation_[k] = T.data[i][0];
-      } else {
-        DALI_ENFORCE(T.shape[0][0] == output_pt_dim_, make_string("Expected ",
-          output_pt_dim_, "-D translation vectors, got ", T.shape[0][0], "-D"));
-        for (int i = 0, k = 0; i < N; i++)
-          for (int j = 0; j < output_pt_dim_; j++, k++)
-            per_sample_translation_[k] = T.data[i][j];
-      }
-    } else {
-      translation_.resize(output_pt_dim_, 0.0f);
-      per_sample_translation_.resize(N * output_pt_dim_, 0.0f);
-    }
-  }
-
-  template <typename OutRange, typename InRange>
-  void Repeat(OutRange &&out, InRange &&in, int times) {
-    ssize_t n = size(in);
-    resize_if_possible(out, n * times);
-    for (ssize_t i = 0, k = 0; i < times; i++)
-      for (ssize_t j = 0; j < n; j++, k++)
-        out[k] = in[j];
   }
 
  private:
